pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dsbench-notebook-agent"
version = "0.1.0"
description = "Notebook-persistent ReACT agent for DSBench-style ML workflows."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "ipykernel>=6.29.5",
  "jupyter_client>=8.6.3",
  "nbclient>=0.10.2",
  "nbformat>=5.10.4",
  "numpy>=2.2.3",
  "openai>=1.68.2",
  "pandas>=2.2.3",
  "python-dotenv>=1.0.1",
  "scikit-learn>=1.6.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.5",
]

[tool.setuptools]
py-modules = [
  "agent",
  "environment",
  "main",
  "tools",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
